requirements.txt
aiogram
groq
google-generativeai
